Added documentation to code and added additional functionality to the package and bumped the version

+++ mystery.go
@@ -5,6 +5,12 @@ import (
 	"strconv"
 )
 
+// Simplify takes a uint64 and recursively (up to 9 times) reduces a number like 1776 to 3.
+//
+//	Example: 1776 = 3
+//	      1+7+7 = 15 becomes (15)6 = 3 ; becomes (1+5=6)6 = 3 ; becomes 6+6 = 12 becomes 1+2 becomes 3
+//
+// Therefore: Simplify(1776) = 3
 func Simplify(in uint64) (out uint64, err error) {
 	run := func(in uint64) (out uint64, err error) {
 		chars := []byte(strconv.FormatUint(in, 10))
@@ -36,11 +42,12 @@ func Simplify(in uint64) (out uint64, err error) {
 	return out, err
 }
 
+// Majestic determines if the in uint64 digits are only 3, 6 or 9
 func Majestic(in uint64, err error) (out bool) {
 	if err != nil {
 		return false
 	}
-	chars := []byte(strconv.FormatUint(uint64(in), 10))
+	chars := []byte(strconv.FormatUint(in, 10))
 	for _, c := range chars {
 		if c != '3' && c != '6' && c != '9' {
 			return false
